requirements.txt
aiohttp>=3.9.0
discord.py>=2.3.2
openai>=1.12.0
python-dotenv>=1.0.1
python-telegram-bot>=20.7
